requirements.txt

discord.py==2.5.2
yt-dlp
aiohttp
python-dotenv
